# rust/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Image Trace RS API
  version: 3.0.0
  description: |
    图像比对与文档图片提取服务（Rust 重写版）。

    核心能力：
    - 项目 / 图片 / 文档管理
    - 上传图片或 PDF、DOCX、PPTX 文档并自动提取其中的图片
    - 多算法相似度比对：感知哈希（phash/dhash/ahash/whash/colorhash）、
      像素级（ssim/histogram/template）、局部特征描述子（orb/akaze/sift）
    - 智能查重：多算法投票 + 并查集聚类，识别旋转/翻转/裁剪/切片重组后的副本
    - 相似度矩阵、特征匹配详情、可视化与重复图片报告

    约定：
    - 所有路径以 `/v1` 为前缀
    - 时间均为 RFC 3339 / ISO 8601 UTC 字符串
    - 相似度得分均为 [0, 1] 浮点数，1 表示完全相同
    - 文件相关字段 `file_path` 为相对数据目录的路径，可通过 `/v1/files/{path}` 下载
servers:
  - url: http://127.0.0.1:8000
    description: 本地默认实例
...
paths:
...
  /v1/projects:
    get:
      tags: [projects]
      summary: 项目列表
      operationId: listProjects
      parameters:
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 项目列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
    post:
      tags: [projects]
      summary: 创建项目
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: 项目已创建
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
  /v1/projects/{projectId}:
    get:
      tags: [projects]
      summary: 获取项目详情
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 项目详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [projects]
      summary: 删除项目（连同图片与分析记录）
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 已删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/projects/{projectId}/images:
    get:
      tags: [images]
      summary: 项目内图片列表
      operationId: listProjectImages
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 图片列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Image'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/images/{imageId}:
    delete:
      tags: [images]
      summary: 删除图片（清理缓存与特征）
      operationId: deleteImage
      parameters:
        - $ref: '#/components/parameters/ImageId'
      responses:
        '200':
          description: 已删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/images/{imageId}/thumbnail:
    get:
      tags: [images]
      summary: 获取图片缩略图
      operationId: getThumbnail
      parameters:
        - $ref: '#/components/parameters/ImageId'
        - name: size
          in: query
          schema: { type: integer, default: 400, minimum: 16, maximum: 2048 }
      responses:
        '200':
          description: JPEG 缩略图
          content:
            image/jpeg: {}
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/upload:
    post:
      tags: [upload]
      summary: 上传文件（图片或文档）
      description: |
        multipart 上传。图片直接入库并触发特征预计算；
        PDF/DOCX/PPTX 文档自动抽取内嵌图片，抽取出的每张图同样入库并预计算。
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [project_id, file]
              properties:
                project_id: { type: integer }
                file: { type: string, format: binary }
      responses:
        '200':
          description: 处理结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: 文档解析失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/projects/{projectId}/feature-status:
    get:
      tags: [analysis]
      summary: 特征预计算进度
      operationId: featureStatus
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 特征状态
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureStatus'
  /v1/projects/{projectId}/features/recompute:
    post:
      tags: [analysis]
      summary: 重新触发未完成图片的特征预计算
      operationId: recomputeFeatures
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: 已触发数量
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecomputeResult'
  /v1/projects/{projectId}/compare:
    post:
      tags: [analysis]
      summary: 执行单算法相似度比对
      description: |
        对项目内全部图片两两比对，按阈值进行并查集聚类，返回相似分组与独立图片。
        `rotation_invariant=true` 时会对每张图生成 8 个方向变体（4 旋转 × 2 翻转）取最高分。
      operationId: compareProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompareRequest'
      responses:
        '200':
          description: 比对结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparisonResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/projects/{projectId}/smart-compare:
    post:
      tags: [analysis]
      summary: 智能查重（多算法投票）
      description: |
        使用全部可用算法并行计算 N×N 相似度矩阵，对超过阈值的图片对累计投票；
        票数 ≥ `min_agree` 且至少命中一种哈希算法时才确认为重复对，
        最后通过并查集生成重复分组。对旋转/翻转/轻微裁剪鲁棒。
      operationId: smartCompare
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SmartCompareRequest'
      responses:
        '200':
          description: 智能查重结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SmartCompareResult'
  /v1/projects/{projectId}/matrix:
    get:
      tags: [analysis]
      summary: N×N 相似度矩阵
      operationId: pairwiseMatrix
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: algorithm
          in: query
          schema: { $ref: '#/components/schemas/Algorithm' }
        - name: rotation_invariant
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: 相似度矩阵
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatrixResult'
  /v1/projects/{projectId}/report:
    get:
      tags: [analysis]
      summary: 重复图片分析报告
      operationId: projectReport
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: algorithm
          in: query
          schema: { $ref: '#/components/schemas/Algorithm' }
        - name: threshold
          in: query
          schema: { type: number, default: 0.85 }
        - name: rotation_invariant
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: 完整报告
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
  /v1/analysis-runs:
    get:
      tags: [analysis]
      summary: 分析运行历史
      operationId: listAnalysisRuns
      parameters:
        - name: project_id
          in: query
          required: true
          schema: { type: integer }
        - $ref: '#/components/parameters/Skip'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 运行记录
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnalysisRun'
  /v1/analysis-runs/{runId}:
    get:
      tags: [analysis]
      summary: 单次分析结果详情
      operationId: getAnalysisRun
      parameters:
        - name: runId
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 运行详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisRunDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/match/pairs:
    post:
      tags: [analysis]
      summary: 两图特征点匹配详情
      description: 返回特征点坐标与匹配连线数据，用于前端可视化。
      operationId: matchData
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_a_id, image_b_id]
              properties:
                image_a_id: { type: integer }
                image_b_id: { type: integer }
                algorithm:
                  allOf:
                    - $ref: '#/components/schemas/Algorithm'
                  description: 仅限描述子类算法（orb/akaze/sift）
      responses:
        '200':
          description: 匹配数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchData'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/match/visualize:
    post:
      tags: [analysis]
      summary: 生成两图特征匹配连线图
      operationId: visualizeMatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_a_id, image_b_id]
              properties:
                image_a_id: { type: integer }
                image_b_id: { type: integer }
                algorithm:
                  allOf:
                    - $ref: '#/components/schemas/Algorithm'
      responses:
        '200':
          description: 生成的可视化文件信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisualizationResult'
  /v1/match/slices:
    post:
      tags: [analysis]
      summary: 切片/子图检测
      description: |
        将图 B 均匀切成 R×C 个切片，每个切片分别与图 A 做局部匹配。
        用于识别「整图被切片打散」「大图嵌入小图/小图被放大嵌入」等变换。
      operationId: sliceMatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image_a_id, image_b_id]
              properties:
                image_a_id: { type: integer }
                image_b_id: { type: integer }
                rows: { type: integer, default: 2, minimum: 1, maximum: 8 }
                cols: { type: integer, default: 2, minimum: 1, maximum: 8 }
      responses:
        '200':
          description: 每个切片的匹配得分
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SliceMatchResult'
  /v1/files/{path}:
    get:
      tags: [files]
      summary: 下载数据目录内文件
      operationId: downloadFile
      parameters:
        - name: path
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 文件内容
          content:
            application/octet-stream: {}
        '403':
          description: 路径越界
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: integer }
    ImageId:
      name: imageId
      in: path
      required: true
      schema: { type: integer }
    Skip:
      name: skip
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 100, minimum: 1, maximum: 500 }
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        detail: { type: string }
    Message:
      type: object
      properties:
        message: { type: string }
...
    Project:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string, nullable: true }
        image_count: { type: integer }
        created_at: { type: string, format: date-time }
    ProjectCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 255 }
        description: { type: string, maxLength: 1000 }
...
    RecomputeResult:
      type: object
      properties:
        triggered: { type: integer }
        message: { type: string }
...
    VisualizationResult:
      type: object
      properties:
        file_path: { type: string }
        media_type: { type: string }
...